:liptsick: extract `WindowTitle` type into its own file, some renames
